fix fmt_fp.c(329) : warning C4244: 'return' : conversion from 'ptrdiff_t' to 'int', possible loss of data

+++ src/fmt_fp.c
@@ -326,7 +326,7 @@ fmt_fp(struct fmt_args *f, long double y, ptrdiff_t p, uint8_t fl, int t)
 
   pad(f, ' ', 0, pl+l, fl^LEFT_ADJ);
 
-  return pl+l;
+  return (int)pl+(int)l;
 }
 
 static int
